+++ reservation/app/src/main/res/layout/details_fragment.xml
@@ -13,11 +13,15 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:id="@+id/selectedTitle"
-        android:textSize="32sp"
-        android:background="#2ECC71"
-        android:layout_weight="1"
-        android:gravity="center"
-        android:text="TextView for TITLE"/>
+        android:text="TextView for TITLE"
+        android:textAppearance="?android:attr/textAppearanceLarge" />/>
+
+    <TextView
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:id="@+id/period"
+        android:text="캠핑 기간"
+        android:textAppearance="?android:attr/textAppearanceLarge" />/>
 
     <TextView
         android:id="@+id/textView1"
@@ -31,31 +35,56 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:checked="true"
-        android:text="A구역 13,000원" />
+        android:text="가족텐트 33,000원" />
 
     <CheckBox
         android:id="@+id/checkBox2"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="B구역 10,000원" />
+        android:text="몽골텐트_대 50,000원" />
 
     <CheckBox
         android:id="@+id/checkBox3"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="C구역 13,000원" />
+        android:text="몽골텐트_대(냉_난방포함) 68,000원" />
 
     <CheckBox
         android:id="@+id/checkBox4"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="D구역 13,000" />
+        android:text="몽골텐트_중 40,000원" />
 
-    <Button
-        android:id="@+id/button1"
+    <CheckBox
+        android:id="@+id/checkBox5"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="선택완료" />
+        android:text="몽골텐트_중(냉_난방포함) 58,000원" />
+
+    <CheckBox
+        android:id="@+id/checkBox6"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="몽골텐트_특대(5x4) 70,000원" />
+
+    <CheckBox
+        android:id="@+id/checkBox7"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="몽골텐트_특대(5x5) 70,000원" />
+
+    <CheckBox
+        android:id="@+id/checkBox8"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="캐빈텐트 44,500원" />
+
+    <CheckBox
+        android:id="@+id/checkBox9"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="캐빈텐트(냉_난방 포함) 62,500원" />
+
     <TextView
         android:id="@+id/textView2"
         android:layout_width="wrap_content"
@@ -63,10 +92,25 @@
         android:text="선택결과창"
         android:textAppearance="?android:attr/textAppearanceLarge" />
 
-    <Spinner
-        android:id="@+id/txt_question_type"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:gravity="right"></Spinner>
+    <RelativeLayout
+        android:id="@+id/rl_done_btn"
+        android:layout_width="270dp"
+        android:layout_height="48dp"
+        android:background="@color/colorTest"
+        android:layout_marginBottom="17dp"
+        android:layout_alignParentBottom="true"
+        android:layout_centerHorizontal="true">
+
+        <TextView
+            android:text="예약"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_centerVertical="true"
+            android:textSize="16sp"
+            android:textColor="#ffffff"
+            android:layout_centerHorizontal="true"
+            android:id="@+id/button1"
+            />
+    </RelativeLayout>
 
 </LinearLayout>
